Code refactor,Pact, Functional Test cases
