+++ src/main/java/uk/gov/hmcts/reform/cwrdapi/util/ValidateStaffProfileChildren.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.cwrdapi.util;
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
+import javax.validation.Constraint;
 import javax.validation.Payload;
 
 import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
@@ -11,6 +12,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Target({TYPE, ANNOTATION_TYPE})
 @Retention(RUNTIME)
+@Constraint(validatedBy = {StaffProfileChildListValidator.class})
 @Documented
 public @interface ValidateStaffProfileChildren {
     String message() default "";
